chore: flutter analyze
