+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/board/board_node_widget.dart
@@ -16,6 +16,7 @@ class BoardBlockComponentBuilder extends BlockComponentBuilder {
     this.configuration = const BlockComponentConfiguration(),
   });
 
+  @override
   final BlockComponentConfiguration configuration;
 
   @override
